+++ internal/clients/state/v3/state.go
@@ -3,7 +3,6 @@ package state
 import (
 	"bytes"
 	"context"
-	"fmt"
 	"io"
 	"sync"
 
@@ -71,9 +70,12 @@ func NewClient(ctx context.Context, pbClient pb.PluginClient, tableName string) 
 	}); err != nil {
 		return nil, err
 	}
+	if _, err := writeClient.CloseAndRecv(); err != nil {
+		return nil, err
+	}
 
-	syncClient, err := c.client.Sync(ctx, &pb.Sync_Request{
-		Tables: []string{tableName},
+	readClient, err := c.client.Read(ctx, &pb.Read_Request{
+		Table: tableBytes,
 	})
 	if err != nil {
 		return nil, err
@@ -81,21 +83,14 @@ func NewClient(ctx context.Context, pbClient pb.PluginClient, tableName string) 
 	c.mutex.Lock()
 	defer c.mutex.Unlock()
 	for {
-		res, err := syncClient.Recv()
+		res, err := readClient.Recv()
 		if err != nil {
 			if err == io.EOF {
 				break
 			}
 			return nil, err
 		}
-		var insertMessage *pb.Sync_Response_Insert
-		switch m := res.Message.(type) {
-		case *pb.Sync_Response_MigrateTable:
-			continue
-		case *pb.Sync_Response_Insert:
-			insertMessage = m
-		}
-		rdr, err := ipc.NewReader(bytes.NewReader(insertMessage.Insert.Record))
+		rdr, err := ipc.NewReader(bytes.NewReader(res.Record))
 		if err != nil {
 			return nil, err
 		}
@@ -165,5 +160,5 @@ func (c *Client) GetKey(_ context.Context, key string) (string, error) {
 	if val, ok := c.mem[key]; ok {
 		return val, nil
 	}
-	return "", fmt.Errorf("key not found")
+	return "", nil
 }