+++ internal/servers/destination/v1/destinations.go
@@ -39,7 +39,22 @@ func (s *Server) Configure(ctx context.Context, req *pb.Configure_Request) (*pb.
 	if err != nil {
 		return nil, status.Errorf(codes.InvalidArgument, "failed to marshal spec: %v", err)
 	}
-	return &pb.Configure_Response{}, s.Plugin.Init(ctx, pluginSpec)
+	var pluginSpecMap map[string]any
+	if err := json.Unmarshal(pluginSpec, &pluginSpecMap); err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "failed to unmarshal plugin spec: %v", err)
+	}
+	// this is for backward compatibility
+	if s.spec.BatchSize > 0 {
+		pluginSpecMap["batch_size"] = s.spec.BatchSize
+	}
+	if s.spec.BatchSizeBytes > 0 {
+		pluginSpecMap["batch_size_bytes"] = s.spec.BatchSizeBytes
+	}
+	pluginSpec, err = json.Marshal(pluginSpecMap)
+	if err != nil {
+		return nil, status.Errorf(codes.InvalidArgument, "failed to marshal spec: %v", err)
+	}
+	return &pb.Configure_Response{}, s.Plugin.Init(ctx, pluginSpec, plugin.NewClientOptions{})
 }
 
 func (s *Server) GetName(context.Context, *pb.GetName_Request) (*pb.GetName_Response, error) {
@@ -211,7 +226,7 @@ func (s *Server) DeleteStale(ctx context.Context, req *pb.DeleteStale_Request) (
 		bldr.Field(table.Columns.Index(schema.CqSourceNameColumn.Name)).(*array.StringBuilder).Append(req.Source)
 		bldr.Field(table.Columns.Index(schema.CqSyncTimeColumn.Name)).(*array.TimestampBuilder).AppendTime(req.Timestamp.AsTime())
 		msgs <- &message.WriteDeleteStale{
-			Table:      table,
+			TableName:  table.Name,
 			SourceName: req.Source,
 			SyncTime:   req.Timestamp.AsTime(),
 		}