+++ state/state.go
@@ -14,6 +14,7 @@ import (
 type Client interface {
 	SetKey(ctx context.Context, key string, value string) error
 	GetKey(ctx context.Context, key string) (string, error)
+	Flush(ctx context.Context) error
 }
 
 func NewClient(ctx context.Context, conn *grpc.ClientConn, tableName string) (Client, error) {
@@ -27,3 +28,20 @@ func NewClient(ctx context.Context, conn *grpc.ClientConn, tableName string) (Cl
 	}
 	return nil, fmt.Errorf("please upgrade your state backend plugin. state supporting version 3 plugin has %v", versions.Versions)
 }
+
+type NoOpClient struct{}
+
+func (*NoOpClient) SetKey(_ context.Context, _ string, _ string) error {
+	return nil
+}
+
+func (*NoOpClient) GetKey(_ context.Context, _ string) (string, error) {
+	return "", nil
+}
+
+func (*NoOpClient) Flush(_ context.Context) error {
+	return nil
+}
+
+// static check
+var _ Client = (*NoOpClient)(nil)